factorize synonyms and stop words
